perf: use concrete type for Counter, Gauge, and Histogram metrics (#1590)
## Which problem is this PR solving?

Use concrete type as input when recording Counter, Gauge, and Histogram
metrics. This is to prevent allocation on the heap caused by converting
a concrete type to `interface{}`

## Short description of the changes

- Use `float64` as input type for Gauge and Histogram
- Use `int64` as input type for Counter
- Implement a wrapper for libhoney internal metrics since
transmission.Metrics requires input being passed in as `interface{}`

---------

+++ app/app.go
@@ -44,8 +44,8 @@ func (a *App) Start() error {
 		}
 		cfgMetric := config.ConfigHashMetrics(cfgHash)
 		ruleMetric := config.ConfigHashMetrics(rulesHash)
-		a.Metrics.Gauge("config_hash", cfgMetric)
-		a.Metrics.Gauge("rule_config_hash", ruleMetric)
+		a.Metrics.Gauge("config_hash", float64(cfgMetric))
+		a.Metrics.Gauge("rule_config_hash", float64(ruleMetric))
 	}
 
 	a.Logger.Debug().Logf("Starting up App...")
